[INLONG-10784][Sort] Refactor the structure of inlongmsg-rowdata
