+++ inlong-sort/sort-formats/format-rowdata/format-inlongmsg-rowdata-base/src/main/java/org/apache/inlong/sort/formats/inlongmsg/rowdata/AbstractInLongMsgDecodingFormat.java
@@ -15,7 +15,9 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.sort.formats.inlongmsg;
+package org.apache.inlong.sort.formats.inlongmsg.rowdata;
+
+import org.apache.inlong.sort.formats.inlongmsg.InLongMsgMetadata;
 
 import org.apache.flink.api.common.serialization.DeserializationSchema;
 import org.apache.flink.table.connector.format.DecodingFormat;
