+++ inlong-sort/sort-formats/format-rowdata/format-inlongmsg-rowdata-csv/src/main/java/org/apache/inlong/sort/formats/inlongmsgcsv/InLongMsgCsvDecodingFormat.java
@@ -17,7 +17,7 @@
 
 package org.apache.inlong.sort.formats.inlongmsgcsv;
 
-import org.apache.inlong.sort.formats.inlongmsg.AbstractInLongMsgDecodingFormat;
+import org.apache.inlong.sort.formats.inlongmsg.rowdata.AbstractInLongMsgDecodingFormat;
 
 import org.apache.commons.lang3.StringEscapeUtils;
 import org.apache.flink.api.common.serialization.DeserializationSchema;
