+++ inlong-sort/sort-formats/format-rowdata/format-inlongmsg-rowdata-kv/src/main/java/org/apache/inlong/sort/formats/inlongmsgkv/InLongMsgKvDecodingFormat.java
@@ -17,7 +17,7 @@
 
 package org.apache.inlong.sort.formats.inlongmsgkv;
 
-import org.apache.inlong.sort.formats.inlongmsg.AbstractInLongMsgDecodingFormat;
+import org.apache.inlong.sort.formats.inlongmsg.rowdata.AbstractInLongMsgDecodingFormat;
 
 import org.apache.commons.lang3.StringEscapeUtils;
 import org.apache.flink.api.common.serialization.DeserializationSchema;
